Deleted unused files and functions. Refactored some code to appropriate directories.
