+++ app/src/main/java/io/xavier/topwsb/common/Constants.kt
@@ -11,21 +11,13 @@ object Constants {
  * Database constants
  */
 const val TABLE_NAME_TRENDING_STOCKS = "trending_stocks"
-const val TABLE_NAME_STOCK_OVERVIEW = "stock_overview"
 const val TABLE_NAME_COMMENTS = "wsb_comments"
 
 // Common column names
 const val TICKER_STOCK_COL_NAME = "ticker_trending_stock"
-const val TICKER_OVERVIEW_COL_NAME = "ticker_stock_overview"
 
 // Trending stock table column names
 const val LAST_UPDATED_COL_NAME = "last_updated_utc"
 const val NUMBER_OF_COMMENTS_COL_NAME = "number_of_comments"
 const val SENTIMENT_COL_NAME = "sentiment"
 const val SENTIMENT_SCORE_COL_NAME = "sentiment_score"
-
-// Stock overview table column names
-const val COMPANY_NAME_COL_NAME = "company_name"
-const val HIGH_52WEEK_COL_NAME = "52_week_high"
-const val LOW_52WEEK_COL_NAME = "52_week_low"
-const val ANALYST_TGT_COL_NAME = "analyst_target"
