+++ app/src/main/java/io/xavier/topwsb/data/local/TrendiesDatabase.kt
@@ -3,7 +3,7 @@ package io.xavier.topwsb.data.local
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import io.xavier.topwsb.data.local.entities.TrendingStockEntity
-import io.xavier.topwsb.domain.model.WsbComment
+import io.xavier.topwsb.domain.model.wsb_comment.WsbComment
 
 /**
  * Database used to cache trending stock information.
