+++ app/src/main/java/io/xavier/topwsb/domain/model/wsb_comment/WsbComment.kt
@@ -1,4 +1,4 @@
-package io.xavier.topwsb.domain.model
+package io.xavier.topwsb.domain.model.wsb_comment
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
@@ -8,6 +8,7 @@ import androidx.room.PrimaryKey
 import io.xavier.topwsb.common.TABLE_NAME_COMMENTS
 import io.xavier.topwsb.common.TICKER_STOCK_COL_NAME
 import io.xavier.topwsb.data.local.entities.TrendingStockEntity
+import io.xavier.topwsb.data.remote.dto.wsb_comments.WsbCommentDto
 
 /**
  * Data class for a comment on the /r/wallstreetbets subreddit.
@@ -37,4 +38,22 @@ data class WsbComment(
     val createdUtc: Long,
     val author: String,
     val link: String
-)
+) {
+    companion object {
+        /**
+         * Builds a WsbComment object from a data transfer object.
+         *
+         * @param dto [WsbCommentDto] retrieved from remote api call
+         * @param ticker symbol of the stock this comment is in reference to
+         * @return [WsbComment] referencing [ticker] built from [dto]
+         */
+        fun buildFromDto(dto: WsbCommentDto, ticker: String): WsbComment = WsbComment(
+            id = 0,
+            text = dto.body,
+            ticker = ticker,
+            createdUtc = dto.created_utc * 1_000, // convert from seconds to ms
+            author = dto.author,
+            link = dto.permalink
+        )
+    }
+}
